% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{barchartPlot}
\alias{barchartPlot}
\title{create a barchartPlot}
\usage{
barchartPlot(dataSrc, t_title, xlabel, ylabel, xdataName, ydataName)
}
\arguments{
\item{dataSrc}{dataframe data source}

\item{t_title}{graph title}

\item{xlabel}{graph x axis label}

\item{ylabel}{graph y axis label}

\item{xdataName}{column name to use for the X axis}

\item{ydataName}{column name to use for the T axis}
}
\value{
a dataframe containing patient record data in json
}
\description{
create a barchartPlot
}
