% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.r
\name{getSample}
\alias{getSample}
\title{Get a sample of patient IDs from an NHM database that meet given criteria}
\usage{
getSample(
  dbDir,
  versionDateName,
  patient_samples,
  base_itt_patient,
  sampleSize
)
}
\arguments{
\item{dbDir}{the directory to find the database}

\item{versionDateName}{the ID of the patient database}

\item{patient_samples}{existing patient_samples to append to}

\item{base_itt_patient}{A function to test if the patient should be sampled}
}
\value{
a dataframe containing patient record versionDateName and IDs
}
\description{
Get a sample of patient IDs from an NHM database that meet given criteria
}
