% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{importDataBase}
\alias{importDataBase}
\title{import an NHM database into R}
\usage{
importDataBase(dbName)
}
\arguments{
\item{dbName}{the number of workers}
}
\value{
a dataframe containing patient record data in json
}
\description{
import an NHM database into R
}
