% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{importNHMDataBase}
\alias{importNHMDataBase}
\title{import an NHM database into R}
\usage{
importNHMDataBase(dbDir, dbName)
}
\arguments{
\item{dbDir}{the directory to find the database}

\item{dbName}{the ID of the patient database}
}
\value{
a dataframe containing patient record data in json
}
\description{
import an NHM database into R
}
