% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{importNHMDataBaseFiles}
\alias{importNHMDataBaseFiles}
\title{import an NHM database into R}
\usage{
importNHMDataBaseFiles(dbDir, dbName, outputDir = dbDir, chunkSize = 1e+05)
}
\arguments{
\item{dbDir}{the directory to find the database}

\item{dbName}{the ID of the patient database}
}
\value{
a character array containing a zip filename of files pointing to patient record data in json
}
\description{
import an NHM database into R
}
