% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.r
\name{importNHMDataBaseSampleData}
\alias{importNHMDataBaseSampleData}
\title{sample an NHM database for patients that meet given criteria}
\usage{
importNHMDataBaseSampleData(dbDir, versionDateName)
}
\arguments{
\item{dbDir}{the directory to find the database}

\item{versionDateName}{the ID of the patient database}
}
\value{
a dataframe containing patient record versionDateName and IDs
}
\description{
sample an NHM database for patients that meet given criteria
}
