% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapReduce_reduce_fromFilelist}
\alias{mapReduce_reduce_fromFilelist}
\title{Map Reduce Reduce function for file lists}
\usage{
mapReduce_reduce_fromFilelist(fileList, key, functions, summary_vars)
}
\arguments{
\item{fileList}{a file list from mapReduce_map_fromtoFileList}

\item{key}{a c("key") containing the column names to group by}

\item{functions}{a c("functions") to apply to the reduction}

\item{summary_vars}{a c("variables") to apply the functions}
}
\value{
a dataframe with the result of the reduction step
}
\description{
Process a list of mapped dataframes and return a dataframe containing c(key) with c(functions) applied to c(summary_vars)
}
\examples{
molecules <- mapReduce_reduce(list_data,c("molecule"),c("sum"),c("count"))

}
