% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreduce.R
\name{mapReduce_setBatchSize}
\alias{mapReduce_setBatchSize}
\title{Map Reduce set batch size}
\usage{
mapReduce_setBatchSize(batchSize)
}
\arguments{
\item{batchSize}{the number of elements to run in a batch}
}
\value{
void
}
\description{
By default, 35,000 is used update this relative to available memory
}
