% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{multilinePlot}
\alias{multilinePlot}
\title{Create a plot with multiple lines and shapes}
\usage{
multilinePlot(
  dataSrc,
  t_title,
  xlabel,
  ylabel,
  xdataName,
  ydataName,
  yUnits = "",
  seqB = 0,
  seqE = 110,
  seqBy = 10
)
}
\description{
Create a plot with multiple lines and shapes
}
\examples{
d<-data.frame()
d<-rbind(d,data.frame(age=c(18:100),Legend="male",weight=rnorm(83,100,10)))
d<-rbind(d,data.frame(age=c(18:100),Legend="female",weight=rnorm(83,80,5)))
multilinePlot(d,"Random Weight Graph","Age","Weight(kg)",age,weight)

}
