% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHMAnalysis.r
\name{populationPyramid}
\alias{populationPyramid}
\title{Create a Population Pyramid graph}
\usage{
populationPyramid(data, gTitle)
}
\arguments{
\item{data}{data containing the age,sex and population size data : data.frame(age,sex,Population)}

\item{gTitle}{The title to put on the graph}
}
\description{
Create a Population Pyramid graph
}
