% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{rtruncnorm}
\alias{rtruncnorm}
\title{sample normal distribution with min and max}
\usage{
rtruncnorm(n, mean, sd, min = -Inf, max = Inf)
}
\arguments{
\item{n}{number of samples}

\item{mean}{mean of samples}

\item{sd}{Standard Deviation of samples}

\item{min}{min of samples}

\item{max}{max of samples}
}
\value{
Samples with given parameters
}
\description{
sample normal distribution with min and max
}
