% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.r
\name{sampleData}
\alias{sampleData}
\title{sample an NHM database for patients that meet given criteria}
\usage{
sampleData(dbDir, versionDateName, base_itt_patient)
}
\arguments{
\item{dbDir}{the directory to find the database}

\item{versionDateName}{the ID of the patient database}

\item{base_itt_patient}{A function to test if the patient should be sampled}
}
\value{
a dataframe containing patient record versionDateName and IDs
}
\description{
sample an NHM database for patients that meet given criteria
}
