% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHMAnalysis.r
\name{setanaysisParameters}
\alias{setanaysisParameters}
\title{set anaysis parameters to use with getParameters_mapFunction}
\usage{
setanaysisParameters(group, items)
}
\arguments{
\item{group}{type of parameters to extract ("Parameters" or "Tests")}

\item{items}{a vector of parameters to extract}
}
\description{
set anaysis parameters to use with getParameters_mapFunction
}
