% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{SpreadsheetLettersToNumbers}
\alias{SpreadsheetLettersToNumbers}
\title{Convert Excel column letters into numbers}
\usage{
SpreadsheetLettersToNumbers(s)
}
\arguments{
\item{s}{A string of letters to convert (e.g. "AA")}
}
\value{
The column number represented by this string
}
\description{
Convert Excel column letters into numbers
}
