% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardLib.R
\name{SpreadsheetNumbersToLetters}
\alias{SpreadsheetNumbersToLetters}
\title{Convert Excel column numbers into letters}
\usage{
SpreadsheetNumbersToLetters(col, row)
}
\arguments{
\item{col}{the column number (e.g. 702)}

\item{row}{the row number (e.g. 25)}
}
\value{
The string for use in spreadsheet functions (e.g. ZZ25)
}
\description{
Convert Excel column numbers into letters
}
