% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHMAnalysis.r
\name{waterFallGraph}
\alias{waterFallGraph}
\title{seeehttps://stackoverflow.com/questions/48259930/how-to-create-a-stacked-waterfall-chart-in-r}
\usage{
waterFallGraph(
  data,
  graphTitle,
  yTitle,
  xaxis,
  category,
  valueField,
  threshold = 8e+07
)
}
\description{
seeehttps://stackoverflow.com/questions/48259930/how-to-create-a-stacked-waterfall-chart-in-r
}
